Move default configs from BL
