+++ OutOfSchool/OutOfSchool.BackgroundJobs/Config/DefaultQuartzConfig.cs
@@ -1,4 +1,4 @@
-namespace OutOfSchool.BusinessLogic.Config.Quartz;
+namespace OutOfSchool.BackgroundJobs.Config;
 
 public static class DefaultQuartzConfig
 {
