+++ OutOfSchool/OutOfSchool.BackgroundJobs/Config/QuartzConfig.cs
@@ -1,4 +1,4 @@
-namespace OutOfSchool.BusinessLogic.Config.Quartz;
+namespace OutOfSchool.BackgroundJobs.Config;
 
 public class QuartzConfig
 {
