+++ OutOfSchool/OutOfSchool.WebApi/Extensions/EmailSenderExtensions.cs
@@ -1,4 +1,5 @@
-﻿using OutOfSchool.BusinessLogic.Common.QuartzConstants;
+﻿using OutOfSchool.BackgroundJobs.Config;
+using OutOfSchool.BusinessLogic.Common.QuartzConstants;
 using OutOfSchool.EmailSender.Quartz;
 using Quartz;
 using Quartz.Impl.Matchers;
