+++ OutOfSchool/OutOfSchool.WebApi/Extensions/Startup/QuartzExtension.cs
@@ -1,3 +1,4 @@
+using OutOfSchool.BackgroundJobs.Config;
 using Quartz;
 using Quartz.Impl;
