+++ OutOfSchool/OutOfSchool.WebApi/OutOfSchool.WebApi.csproj
@@ -79,6 +79,7 @@
         <PackageReference Include="Microsoft.AspNetCore.HeaderPropagation" />
     </ItemGroup>
     <ItemGroup>
+        <ProjectReference Include="..\OutOfSchool.BackgroundJobs\OutOfSchool.BackgroundJobs.csproj" />
         <ProjectReference Include="..\OutOfSchool.BusinessLogic\OutOfSchool.BusinessLogic.csproj" />
         <ProjectReference Include="..\OutOfSchool.Common\OutOfSchool.Common.csproj" />
         <ProjectReference Include="..\OutOfSchool.DataAccess\OutOfSchool.DataAccess.csproj" />
